+++ HISTORY.md
@@ -1,3 +1,17 @@
+unreleased
+==========
+
+  * Methods are no longer enumerable on `req.session` object
+
+1.13.1 / 2016-01-29
+===================
+
+  * deps: parseurl@~1.3.1
+    - perf: enable strict mode
+  * deps: uid-safe@~2.1.0
+    - Use `random-bytes` for byte source
+  * perf: enable strict mode
+
 1.13.0 / 2016-01-10
 ===================